// src/types/tableTypes.ts
export interface SurveyTask {
    task_id: number;            // Уникальный идентификатор задания
    survey_id: number;          // Идентификатор опроса
    description: string;        // Описание задания
    price: number;              // Оплата за задание
    task_type: survey_task_type_enum;// Тип задания (например, "видео", "фото")
    created_at: string;         // Время создания задания (формат TIMESTAMP)
}
// ...
export type survey_task_type_enum  ='live'| 'photo'| 'video'